Moved sort functionality back into the EntityTreeBrowser. The code now lies in EntityBrowserUtils.
