+++ src/main/java/org/sagebionetworks/web/client/widget/entity/browse/EntityBrowserUtils.java
@@ -1,6 +1,8 @@
 package org.sagebionetworks.web.client.widget.entity.browse;
 
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
 
 import org.sagebionetworks.repo.model.EntityHeader;
@@ -89,4 +91,13 @@ public class EntityBrowserUtils {
 		});
 	}
 
+	public static void sortEntityHeadersByName(List<EntityHeader> list) {
+		Collections.sort(list, new Comparator<EntityHeader>() {
+	        @Override
+	        public int compare(EntityHeader o1, EntityHeader o2) {
+	        	return o1.getName().toLowerCase().compareTo(o2.getName().toLowerCase());
+	        }
+		});
+	}
+	
 }
