+++ src/test/java/org/sagebionetworks/web/unitclient/widget/entity/browse/EntityBrowserUtilsTest.java
@@ -19,6 +19,7 @@ import org.sagebionetworks.schema.adapter.org.json.AdapterFactoryImpl;
 import org.sagebionetworks.web.client.GlobalApplicationState;
 import org.sagebionetworks.web.client.SynapseClientAsync;
 import org.sagebionetworks.web.client.widget.entity.browse.EntityBrowserUtils;
+import org.sagebionetworks.web.client.widget.entity.renderer.EntityListUtil;
 import org.sagebionetworks.web.test.helper.AsyncMockStubber;
 
 import com.google.gwt.user.client.rpc.AsyncCallback;
@@ -56,5 +57,27 @@ public class EntityBrowserUtilsTest {
 		assertTrue(returned.contains(h1));
 		assertTrue(returned.contains(h2));
 		
+	}
+	
+	@Test
+	public void testSortEntityHeadersByName() {
+		EntityHeader header1 = new EntityHeader();
+		EntityHeader header2 = new EntityHeader();
+		EntityHeader header3 = new EntityHeader();
+		header1.setName("Abra");
+		header2.setName("marill");
+		header3.setName("Zubat");
+		
+		List<EntityHeader> list = new ArrayList<EntityHeader>();
+		list.add(header3);
+		list.add(header2);
+		list.add(header1);
+		
+		// [Zubat, marill, Abra] --> [Abra, marill, Zubat]
+		EntityBrowserUtils.sortEntityHeadersByName(list);
+		
+		assertEquals(list.get(0), header1);
+		assertEquals(list.get(1), header2);
+		assertEquals(list.get(2), header3);
 	}
 }
